test: delete proven duplicates and consolidate split files
Deletions are limited to tests proven redundant by comparing normalised
ASTs, then gated on a per-file uncovered-line diff. Eight go:

- tests/test_pipeline/test_handlers.py carried a TestSolarHandler class
  whose four tests were all verbatim copies of test_solar_handler.py,
  which covers the same ground plus eleven more cases. Class removed.
- The same file's TestClimateHandler duplicated three of four cases from
  test_climate_handler.py::TestClimateHandlerGating. Kept the one that is
  unique here (no climate readings), dropped the rest.
- The DEFAULT_TRANSIT_TIMEOUT_SECONDS == 45 lock existed identically in
  test_issue_172 and test_issue_271. Kept #271's, which sits alongside
  the other two transit-timeout constant checks.

Two candidates that looked like duplicates are NOT, and both stay:

- test_handlers.py's DefaultHandler describe_skip test has a body
  byte-identical to test_weather_handler.py's, but `self.handler` differs
  -- they cover different handlers.
- test_is_motion_detected_media_player_states appears in
  test_managers/test_motion.py and test_motion_control.py with the same 9
  cases, but one drives MotionManager.is_motion_detected directly and the
  other goes through the coordinator property that delegates to it.

Consolidation, no tests lost:

- Four basenames existed twice under tests/, which broke grep and made
  node IDs ambiguous. Renamed the subpackage copy of each: test_venetian,
  test_louvered_roof, test_cloud_suppression, test_helpers. No duplicate
  basename remains anywhere in tests/.
- Issue #293 was split across two files covering the same auto_control=OFF
  defect pair, and the second rebuilt inline the coordinator stub the
  first had a helper for. Merged into test_issue_293_auto_off.py.
- test_coordinator_apply_user_axis.py (3 tests) and
  test_coordinator_axis_discovery.py (4) are both issue #725 axis work;
  merged into test_coordinator_axes.py, disambiguating the two _coord()
  helpers that shared a name.
- test_glare_zone_handler_fix.py folded into its neighbour: the `_fix`
  suffix said nothing a year on, and its _make_vertical_cover was a copy
  of the neighbour's differing only by docstring.
- test_config_flow_proxy_cover.py and test_config_flow_outside_temp_source.py
  (3 and 2 tests) fold into test_config_flow_schema_placement.py, which
  exists for exactly this -- asserting which key lives on which step.
- test_pipeline_types_venetian.py sat at tests/ root while testing
  pipeline/types.py; moved into tests/test_pipeline/.

Issue #756's two files are left alone: they share an issue number but
test different subsystems (venetian deferred tilt vs coordinator dispatch),
and merging would force neutralize_venetian_delays onto 12 unrelated tests.

9,122 -> 9,114 collected, exactly the 8 deletions. Per-file line coverage
unchanged: no line became uncovered, line-rate 0.9753 both sides.
